[lib/core] Fix bug related to |UpdateRex()|.
|UpdateRex| takes an Operand as an argument and an optional boolean
flag, |Patch_Opcode|. When a |Register| operand is passed to
|UpdateRex|, the initial step involves checking whether the `r` bit is set.
If it's not set, we then proceed to set it and immediately return.
However, If the `r` bit is initially set, we intend to set the `b` bit
instead. Howerver, this logic functions correctly only when the 3-lower bits
of the opcode do not encode a register. In such cases, we disregard the
`r` bit and set the `b` bit accordingly.